refactor project migration, add api for get list migrations
